amat_rick_and_morty/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.5
django-rest-swagger==2.2.0
djangorestframework==3.13.1
docopt==0.6.2
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
openapi-codec==1.3.2
pandas==1.4.2
pipreqs==0.4.11
pydantic==1.9.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
yarg==0.1.9
